+++ lib/quote/quote_view_mobile.dart
@@ -276,7 +276,7 @@ class _QuantityWidgetState extends State<_QuantityWidget> {
     textEditingController.addListener(() {
       if(indicator == true) {
         _model.onUpdateQuantity(
-          widget.i, widget.b, int.parse(textEditingController.text),);
+          widget.i, widget.b, double.parse(textEditingController.text),);
         widget.listenerUpdateTotals();
       }
       indicator = true;
@@ -294,7 +294,7 @@ class _QuantityWidgetState extends State<_QuantityWidget> {
         controller: textEditingController,
         textAlign: TextAlign.center,
         onChanged: (value) {
-          _model.onUpdateQuantity(widget.i, widget.b, int.parse(value), );
+          _model.onUpdateQuantity(widget.i, widget.b, double.parse(value), );
         },
 
       ) : SelectableText(
